# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Интернет-магазин
  description: |
    API для взаимодействия с сервисами заказов, биллинга и уведомлений.
    
    Система использует единый JWT токен для аутентификации пользователей во всех сервисах.
    Токен, полученный в сервисе заказов, можно использовать для авторизации в сервисе биллинга и в сервисе уведомлений.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: dev@example.com
servers:
  - url: http://localhost:8080
    description: Сервис заказов - Order Service
  - url: http://localhost:8081
    description: Сервис биллинга - Billing Service
  - url: http://localhost:8082
    description: Сервис уведомлений - Notification Service
...
paths:
...
  /api/v1/users:
    post:
      tags:
        - users
      summary: Создание нового пользователя
      description: Создает нового пользователя (публичный эндпоинт)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
  # Заказы
  /api/v1/orders:
    post:
      tags:
        - orders
      summary: Создание нового заказа
      description: Создает новый заказ и списывает средства со счета пользователя
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Некорректные данные заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Недостаточно средств
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{orderId}:
    get:
      tags:
        - orders
      summary: Получение информации о заказе
      description: Возвращает детальную информацию о заказе по его ID
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: ID заказа
          schema:
            type: integer
      responses:
        '200':
          description: Детальная информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOrderResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет доступа к этому заказу
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userId}/orders:
    get:
      tags:
        - orders
      summary: Получение списка заказов пользователя
      description: Возвращает все заказы, принадлежащие указанному пользователю
      operationId: getUserOrders
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
        - name: limit
          in: query
          description: Количество записей на странице
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          description: Смещение для пагинации
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список заказов пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrdersResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
  # Биллинг
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Ошибка при обработке запроса"
          
    # Схемы для аутентификации
...
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: "user123"
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securePassword123"
    CreateUserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "user123"
        email:
          type: string
          format: email
          example: "user@example.com"
          
    # Схемы для заказов
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 42
        name:
          type: string
          example: "Смартфон X1"
        price:
          type: number
          format: float
          example: 599.99
        quantity:
          type: integer
          example: 1
    CreateOrderRequest:
      type: object
      required:
        - items
        - amount
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        amount:
          type: number
          format: float
          example: 599.99
    CreateOrderResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 5
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        amount:
          type: number
          format: float
          example: 599.99
        status:
          type: string
          enum: [created, paid, shipped, delivered, canceled, pending, failed, completed]
          example: "completed"
        created_at:
          type: string
          format: date-time
    GetOrderResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 5
        amount:
          type: number
          format: float
          example: 599.99
        status:
          type: string
          enum: [created, paid, shipped, delivered, canceled, pending, failed, completed]
          example: "completed"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListOrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/GetOrderResponse'
        total:
          type: integer
          example: 10
          
    # Схемы для биллинга
...